labs/AgentEngineDeploy/requirements.txt
uv
google-cloud
cloudpickle==3.1.1
google-cloud-aiplatform[agent_engines,adk,langchain,ag2,llama_index]==1.114.0
a2a-sdk>=0.3.4
pydantic==2.11.7
google-cloud-modelarmor
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
